+++ tools/xdhwebq.current/CLI/xdhwebq.h
@@ -1,0 +1,44 @@
+/*
+	Copyright (C) 2015 Claude SIMON (http://q37.info/contact/).
+
+	This file is part of XDHWebQ.
+
+	XDHWebQ is free software: you can redistribute it and/or
+	modify it under the terms of the GNU Affero General Public License as
+	published by the Free Software Foundation, either version 3 of the
+	License, or (at your option) any later version.
+
+	XDHWebQ is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+	Affero General Public License for more details.
+
+	You should have received a copy of the GNU Affero General Public License
+	along with XDHWebQ. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#ifndef XDHWEBQ_INC_
+# define XDHWEBQ_INC_
+
+# include "scli.h"
+
+# define BEGIN
+// Don't remove what between above 'BEGIN' and below 'END'.
+/* PLACEHOLDER */
+
+# define NAME_MC			"XDHWebQ"
+# define NAME_UC			"XDHWEBQ"
+# define NAME_LC			"xdhwebq"
+# define WEBSITE_URL		"http://q37.info/tools/xdhwebq/"
+# define AUTHOR_NAME		"Claude SIMON"
+# define AUTHOR_CONTACT		"http://q37.info/contact/"
+# define OWNER_NAME			"Claude SIMON"
+# define OWNER_CONTACT		"http://q37.info/contact/"
+# define COPYRIGHT			COPYRIGHT_YEARS " " OWNER_NAME " (" OWNER_CONTACT ")"	
+
+# define END
+
+SCLI_DEC( xdhwebq )
+
+#endif
+