+++ tools/xdhwebq.current/JRE/xdhwebqjre.h
@@ -1,0 +1,33 @@
+/*
+	Copyright (C) 2015 Claude SIMON (http://q37.info/contact/).
+
+	This file is part of XDHWebQ.
+
+	XDHWebQ is free software: you can redistribute it and/or
+	modify it under the terms of the GNU Affero General Public License as
+	published by the Free Software Foundation, either version 3 of the
+	License, or (at your option) any later version.
+
+	XDHWebQ is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+	Affero General Public License for more details.
+
+	You should have received a copy of the GNU Affero General Public License
+	along with XDHWebQ. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#ifndef XDHWEBQJRE_INC_
+# define XDHWEBQJRE_INC_
+
+# define NAME_MC			"XDHWebQJRE"
+# define NAME_LC			"xdhwebqjre"
+# define NAME_UC			"XDHWEBQJRE"
+# define WEBSITE_URL		"http://q37.info/"
+# define AUTHOR_NAME		"Claude SIMON"
+# define AUTHOR_CONTACT		"http://q37.info/contact/"
+# define OWNER_NAME			"Claude SIMON"
+# define OWNER_CONTACT		"http://q37.info/contact/"
+# define COPYRIGHT			COPYRIGHT_YEARS " " OWNER_NAME " (" OWNER_CONTACT ")"
+
+#endif