+++ tools/xdhwebq.current/xdwrgstry.cpp
@@ -1,0 +1,28 @@
+/*
+	Copyright (C) 2015 Claude SIMON (http://q37.info/contact/).
+
+	This file is part of XDHWebQ.
+
+	XDHWebQ is free software: you can redistribute it and/or
+	modify it under the terms of the GNU Affero General Public License as
+	published by the Free Software Foundation, either version 3 of the
+	License, or (at your option) any later version.
+
+	XDHWebQ is distributed in the hope that it will be useful,
+	but WITHOUT ANY WARRANTY; without even the implied warranty of
+	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+	Affero General Public License for more details.
+
+	You should have received a copy of the GNU Affero General Public License
+	along with XDHWebQ. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#include "xdwrgstry.h"
+
+using namespace xdwrgstry;
+
+entry___ xdwrgstry::parameter::ModuleFilename( "ModuleFilename", sclrgstry::Parameters );
+entry___ xdwrgstry::parameter::script::Fallback( "Fallback", xdwrgstry::Scripts );
+
+
+